requirements.txt
rich
pynput
pywin32
keyboard
opencv-python
numpy
pillow
PyDirectInput
nuitka
requests